requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
contourpy==1.1.0
cycler==0.11.0
faiss-cpu==1.7.4
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.41.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
memory-profiler==0.61.0
multidict==6.0.4
numpy==1.24.4
openai==0.27.8
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
six==1.16.0
tiktoken==0.4.0
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.0
